Fixture flavien

+++ front/cicero2/src/Pages/Clients.tsx
@@ -5,6 +5,8 @@ import { NavLink } from 'react-router-dom';
 import { Filesystem, Directory } from "@capacitor/filesystem";
 import { Client } from "../Modele/metier/Client";
 import { Case } from "../Modele/metier/Case";
+import { Button } from '@mui/material';
+import { confirmAlert } from 'react-confirm-alert';
 import Box from "@mui/material/Box";
 import SideBar from '../Components/SideBar';
 import DAOFactory from "../Modele/dao/factory/DAOFactory";
@@ -13,9 +15,9 @@ import SearchIcon from '@mui/icons-material/Search';
 import DeleteIcon from '@mui/icons-material/Delete';
 import NoteAltIcon from '@mui/icons-material/NoteAlt';
 import ClientModal from './Modal/ClientModal';
-import InfoIcon from '@mui/icons-material/Info';
-import { Link } from 'react-router-dom';
-import CloseIcon from '@mui/icons-material/Close';
+import InfoIcon from '@mui/icons-material/Info'; 
+import '../Styles/alert.css';
+import 'react-confirm-alert/src/react-confirm-alert.css';
 
 const searchContainer = {
     display: "flex",
@@ -33,7 +35,8 @@ const searchInput = {
     margin: "5px",
 };
 const StyleAll = {
-    width : '100%'
+    width : '100%', 
+    height: '100%'
 }
 const styletable = {
     border:'2px solid black',
@@ -60,9 +63,19 @@ const defaultCase: Case[] | (() => Case[]) = []
 export default function Clients(){
     const [clientsList, setClientsList] = React.useState(defaultClient);
     const [casesList, setCasesList] = React.useState(defaultCase);
+    const [windowSize, setWindowSize] = React.useState(window.innerWidth);
     const [modalOpen, setModalOpen] = useState(false);
     const [filter, setFilter] = useState("");
     const daoF = DAOFactory.getDAOFactory();
+
+    React.useEffect(() => {
+     function handleResize() {
+         setWindowSize(window.innerWidth);
+       }
+ 
+       window.addEventListener("resize", handleResize);
+     return () => window.removeEventListener("resize", handleResize);
+    }, []);
 
     useEffect (() => {
         async function fetchData() {
@@ -102,8 +115,40 @@ export default function Clients(){
       };
     // Suppression d'un client //
     const deleteClient = async (id: number) => {
-      daoF!.getClientDAO().delete(id);
-      setClientsList(clientsList.filter(c => c.id !== id));
+        confirmAlert({
+            customUI: ({ onClose }) => {
+              return (
+                <div className='custom-ui'>
+                    <h1>Etes-vous sur ?</h1>
+                    <p>Etes-vous sur de vouloir supprimer ce client ?</p>
+                    <Button     
+                        style={{
+                        borderRadius: 5,
+                        backgroundColor: "#d9534f",
+                        padding: "12px 24px",
+                        fontSize: "12px",
+                        color: "white",
+                        marginRight: "5px",
+                        }} 
+                        onClick={onClose}>Annuler</Button>
+                    <Button
+                        style={{
+                        borderRadius: 5,
+                        backgroundColor: "#0275d8",
+                        padding: "12px 24px",
+                        fontSize: "12px",
+                        color: "white",
+                        marginLeft: "5px",
+                        }}
+                        onClick={() => {
+                            daoF!.getClientDAO().delete(id);
+                            setClientsList(clientsList.filter(c => c.id !== id));
+                            onClose();
+                        }}>Confirmer</Button>
+                </div>
+              );
+            }
+        });
     };
 
     const handleSearchChange = (e:any) => {
@@ -111,22 +156,21 @@ export default function Clients(){
       };
 
     const getClientCases = (id: number) => {
-        console.log(casesList);
         if (casesList.length === 0) {
             return " / ";
         }
         let clientCases = casesList.map(c => c.clients.map(cl => cl.id === id ? c : null));
-        console.log(clientCases, "clientCases", id);
-        
         let concat = "";
         for(let i = 0; i < clientCases.length; i++){
-            if(i + 1 === clientCases.length){
-                if(clientCases[i][0] !== null){
-                    concat += clientCases[i][0]!.code.toString();
-                }
-            } else {
-                if(clientCases[i][0] !== null){
-                    concat += clientCases[i][0]!.code.toString() + " - ";
+            for(let y = 0; y < clientCases[i].length; y++){
+                if(i + 1 === clientCases.length){
+                    if(clientCases[i][y] !== null){
+                        concat += clientCases[i][y]!.code.toString();
+                    }
+                } else {
+                    if(clientCases[i][y] !== null){
+                        concat += clientCases[i][y]!.code.toString() + " - ";
+                    }
                 }
             }
         }
@@ -137,109 +181,90 @@ export default function Clients(){
         }
     }
 
+    const checkFilter = (code: string, client: Client) => {
+        if(client.firstname.toLowerCase().includes(filter.toLowerCase()) || client.lastname.toLowerCase().includes(filter.toLowerCase()) || code.toLowerCase().includes(filter.toLowerCase())) {
+            return true;
+        } 
+        return false;
+    }
 
     return (
         <Grid sx={StyleAll}>
-                <Header/>
-                        <button
-            onClick={() => {
-                writeClientFile()
-            }}
-        >
-            Write client
-        </button>
-        <button
-            onClick={() => {
-                readClientFile()
-            }}
-        >
-            Read file
-        </button>
-        <button
-            onClick={() => {
-                deleteClientFile()
-            }}
-        >
-            Delete file
-        </button>
-                <button
-            onClick={() => {
-                deleteClient(2)
-            }}
-        >
-            Delete client
-        </button>
-        <button
-            onClick={() => {
-                updateClientFile()
-            }}
-        >
-            Update client
-        </button>
-                <Box sx={{ display: 'flex', minWidth: 700 }}>
-                    <SideBar />
-                    <main className='main'>
-
-                        <Box maxWidth="lg" sx={MainStyle}>
-                            <Grid sx={{ display: 'flex', justifyContent:'space-between', marginTop:5}}>
+            <Header/>
+            <Box sx={{ display: 'flex', minWidth: 700, height: '100%' }}>
+                <SideBar />
+                <main className='main'>
+                    <Box maxWidth="lg" sx={MainStyle}>
+                        <Grid sx={windowSize >= 750 ? { display: 'flex', justifyContent:'space-between', marginTop:5} : {display:'flex', marginTop:5}}>
+                            <Box sx={{marginLeft:'2%',marginRight:'2%'}}>
                                 <h3>Clients</h3>
-                                <Box sx={{ display: 'flex', justifyContent: 'flex-end'}}>
-                                    <Toolbar>
-                                        <Box sx={searchContainer}>
-                                            <SearchIcon sx={searchIcon} />
-                                            <TextField
-                                            sx={searchInput}
-                                            onChange={handleSearchChange}
-                                            label="Recherche"
-                                            variant="standard"
-                                            />
-                                        </Box>
-                                    </Toolbar>
-                                </Box>
-                            </Grid>
+                            </Box>
+                            <Box sx={windowSize >= 750 ? { display: 'flex', justifyContent: 'flex-end'} : {display:'flex', flexDirection:'column'}}>
+                                <Toolbar>
+                                    <Box sx={searchContainer}>
+                                        <SearchIcon sx={searchIcon} />
+                                        <TextField
+                                        sx={searchInput}
+                                        onChange={handleSearchChange}
+                                        label="Recherche"
+                                        variant="standard"
+                                        />
+                                    </Box>
+                                </Toolbar>
+                            </Box>
+                        </Grid>
+                    </Box>
+                    <Table aria-label="customized table" sx={styletable}>
+                        <TableHead>
+                        <TableRow>
+                            <TableCell align="center">Nom</TableCell>
+                            <TableCell align="center">Affaires associées</TableCell>
+                            <TableCell align="center">Actions</TableCell>
+                        </TableRow>
+                        </TableHead>
+                        <TableBody>
+                            {clientsList.map(client => { 
+                                if (checkFilter(getClientCases(client.id), client)) {
+                                    return <TableRow key={client.id}>
+                                                <TableCell component="th" scope="row" align="center" width={'15%'} >
+                                                        {client.firstname} {client.lastname}
+                                                </TableCell>
+                                                <TableCell align="center" sx={StyleCell}>
+                                                    {getClientCases(client.id)}
+                                                </TableCell>
+                                                <TableCell align="center" width={'15%'} sx={StyleCell}>
+                                                    <InfoIcon color="primary"/>
+                                                    <NoteAltIcon onClick={()=>{ setModalOpen(true) }} color="success"/>
+                                                    {getClientCases(client.id) === " / " ? <DeleteIcon onClick={() => { deleteClient(client.id) }} color="error"/> : <DeleteIcon color="disabled"/> }                   
+                                                </TableCell>
+                                            </TableRow>
+                                }
+                            })}
+                        </TableBody>   
+                    </Table>
+                    <ClientModal modalOpen={modalOpen}>  
+                        <Box maxWidth="lg" sx={ModalStyle}>
+                            <button type="button" className="btn_modalContent" onClick={()=>
+                                {setModalOpen(false);}}> X </button>
+                            <p> Test Modal </p>
                         </Box>
-                            <Table aria-label="customized table" sx={styletable}>
-                                <TableHead>
-                                <TableRow>
-                                    <TableCell align="center">Nom</TableCell>
-                                    <TableCell align="center">Affaires associées</TableCell>
-                                    <TableCell align="center">Actions</TableCell>
-                                </TableRow>
-                                </TableHead>
-                                <TableBody>
-                                    {clientsList.map(client => { 
-                                        if (client.firstname.toLowerCase().includes(filter.toLowerCase()) || client.lastname.toLowerCase().includes(filter.toLowerCase())) {
-                                            return <TableRow key={client.id}>
-                                                        <TableCell component="th" scope="row" align="center" width={'15%'} >
-                                                                {client.firstname} {client.lastname}
-                                                        </TableCell>
-                                                        <TableCell align="center" sx={StyleCell}>
-                                                            {getClientCases(client.id)}
-                                                        </TableCell>
-                                                        <TableCell align="center" width={'15%'} sx={StyleCell}>
-                                                            <Link to={'/clientsInfo?id='+client.id} style={{ textDecoration: 'none' }} > <InfoIcon color="primary"/> </Link>
-                                                            <NoteAltIcon onClick={()=>{ setModalOpen(true) }} color="success"/>
-                                                            <DeleteIcon onClick={() => { deleteClient(client.id) }} color="error"/>                    
-                                                        </TableCell>
-                                                    </TableRow>
-                                            }
-                                            })}
-                                </TableBody>   
-                            </Table>
-                            <ClientModal modalOpen={modalOpen}>  
-                                <Box sx={ModalStyle}>
-                                <Grid sx={{ display: 'flex', justifyContent:'end', marginTop:2, marginRight:2}}>
-                                
-                                    <IconButton onClick={()=> {setModalOpen(false);}}>
-                                        <CloseIcon />
-                                    </IconButton>
-                                    
-                                </Grid>
-                                <p> Modal Client Update </p>
-                                </Box>
-                            </ClientModal>
-                    </main>
-                </Box>
-            </Grid>
+                    </ClientModal>
+                    <button
+                        onClick={() => {
+                            writeClientFile()
+                        }}
+                    >
+                    Write client
+                    </button>
+                    <button
+                        onClick={() => {
+                            deleteClientFile()
+                        }}
+                    >
+                        Delete file
+                    </button>
+                </main>
+            </Box>
+        </Grid>
     );
 }
